Plan table relations are removed since it is removed from db

+++ src/main/java/io/inprice/common/models/Coupon.java
@@ -19,7 +19,7 @@ public class Coupon implements Serializable {
   private Integer days;
   private Long issuedCompanyId;
   private Date issuedAt;
-  private Long planId;
+  private Integer planId;
   private Date createdAt = new Date();
 
 }
